server: split apart apiextensions-apiserver plumbing
